Deprecate Pause termination policy
